# Swagger/swagger.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/transaction":{"get":{"operationId":"TransactionController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"post":{"operationId":"TransactionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}},"/transaction/{id}":{"get":{"operationId":"TransactionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"patch":{"operationId":"TransactionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTransactionDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"delete":{"operationId":"TransactionController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/auth/register":{"post":{"operationId":"AuthController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}}},"/auth/refresh-token":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}}}}},"info":{"title":"Test Plub","description":"API para el test de Plub","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"Name":{"type":"string"},"LastName":{"type":"string"},"Email":{"type":"string"},"Password":{"type":"string"}},"required":["Name","LastName","Email","Password"]},"UpdateUserDto":{"type":"object","properties":{"Name":{"type":"string"},"LastName":{"type":"string"},"Email":{"type":"string"},"Password":{"type":"string"}},"required":["Name","LastName","Email","Password"]},"CreateTransactionDto":{"type":"object","properties":{"Title":{"type":"string"},"Description":{"type":"string"}},"required":["Title","Description"]},"UpdateTransactionDto":{"type":"object","properties":{"Title":{"type":"string"},"Description":{"type":"string"}},"required":["Title","Description"]},"LoginDto":{"type":"object","properties":{"Email":{"type":"string"},"Password":{"type":"string"}},"required":["Email","Password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]}}}}